SAK-51977 Consistency between the 'assignments and resources' tools and their corresponding endpoints - refactor
