+++ assignment/tool/src/webapp/WEB-INF/applicationContext.xml
@@ -28,7 +28,6 @@
         <property name="entityManager" ref="org.sakaiproject.entity.api.EntityManager" />
         <property name="sessionManager" ref="org.sakaiproject.tool.api.SessionManager" />
 		<property name="siteService" ref="org.sakaiproject.site.api.SiteService" />
-		<property name="toolManager" ref="org.sakaiproject.tool.api.ToolManager" />
         <property name="securityService" ref="org.sakaiproject.authz.api.SecurityService" />
         <property name="assignmentSupplementItemService" ref="org.sakaiproject.assignment.api.model.AssignmentSupplementItemService" />
 		<property name="gradingService" ref="org.sakaiproject.grading.api.GradingService" />
